VertretungsplanBot/requirements.txt
requests==2.28.1
discord-py-interactions==4.3.1
python-dateutil==2.8.2
lxml==4.9.1
dacite==1.6.0